Move the SiteConfig to the data-model crate
