+++ crates/data-model/src/lib.rs
@@ -18,6 +18,7 @@ use thiserror::Error;
 
 pub(crate) mod compat;
 pub(crate) mod oauth2;
+mod site_config;
 pub(crate) mod tokens;
 pub(crate) mod upstream_oauth2;
 pub(crate) mod user_agent;
@@ -39,6 +40,7 @@ pub use self::{
         AuthorizationCode, AuthorizationGrant, AuthorizationGrantStage, Client, DeviceCodeGrant,
         DeviceCodeGrantState, InvalidRedirectUriError, JwksOrJwksUri, Pkce, Session, SessionState,
     },
+    site_config::SiteConfig,
     tokens::{
         AccessToken, AccessTokenState, RefreshToken, RefreshTokenState, TokenFormatError, TokenType,
     },
